virtio_net: Fix leaked netdev->refcnt
If after receiving some packets and refilling the queue with buffers,
we detect that more packets are available then we re-schedule the
queue and process them.

This re-scheduling - i.e. calling __netif_rx_schedule() - causes a
netdev reference to be taken.

Once we've finally run out of buffers to process, we return zero and
net_rx_action() drops the reference taken by the original call to
_netif_rx_schedule() in e.g. skb_recv_done().

The reference taken by re-scheduling is always leaked, leading to:

      waiting for eth0 to become free. Usage count = 132568

Fix by immediately dropping the extra reference taken.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>